Fix: Migration script

+++ backend/src/main/resources/db/migration/V1__CREATE-INITIAL-TABLE.sql
@@ -1,7 +1,7 @@
 CREATE TABLE tb_task (
     id SERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
-    cost DOUBLE PRECISION,
+    cost NUMERIC(15,2),
     deadline DATE NOT NULL,
     sorting_position INT UNIQUE NOT NULL
 );
